+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Constellation Changes	
 
+## Changes in August 2022
+
+-   Moved `ClusteringConcept`, `HierarchicalStateAttributeDescription` and
+    `KTrussStateAttributeDescription` from `CoreAlgorithmPlugins` to
+    `CoreAnalyticSchema` so that they could be used in an schema updater to 
+    update the spelling of 'color' in some attributes changed.
+
 ## Changes in July 2022
 
 -   Updated `RestClient` so that `params` are passed as `List<Tuple<String, String>>`
@@ -71,7 +78,10 @@
 
 -   Added `PluginTags` class to hold all tags as constants for `PluginInfo`.
 
--   Added `isRequired` in `PluginParameter` with a getter and a setter, which
+-   Changed the parameter for takeScreenshot in RecentGraphScreenshotUtilities 
+    from filename to filepath.
+
+-   Added `isRequired` in `PluginParameter` with a getter and a setter, which 
     can be used to configure the required plugin parameters to mark as `*required`
     in the swagger.
 
@@ -162,7 +172,7 @@
 
 ## Changes in July 2021
 
--   Update to Quality Control category names and colours in
+-   Update to Quality Control category names and colors in
     `QualityControlEvent` to be easier to understand.
 
 -   Update to `openEdit`, `updateNotesUI`, `createNote` and the constructor in